requirements.txt
numpy
opencv-python
ezkl
torch
Pillow
facenet_pytorch
flask